Tweaks for executing turns
It's still a little wonky, and the sheer amount of console messages used
gets to be a little much. It might be faster to append a div or
something.

The Combatants are choosing and using abilities, but only the HOTs seem
to be applying and messaging. Probably something slightly off.